refactor(core): split candidate-store eager construction seams
Extract path AABB/edge helpers, typed-array indexes/traversal/group
buckets, and spatial shortlist + geometry refine from the 1194-line eager
factory into focused modules. Slim buildCandidateStoreEager to assemble
those pieces and expose CandidateStore methods.

Split the characterization suite by concern (store/cycle/traverse/group/
spatial) with shared fixtures; add unit coverage for path-geometry helpers.
