+++ packages/core/src/candidate-store.ts
@@ -3,7 +3,10 @@
  * - candidate-axis-token.ts — logical axis value identity
  * - candidate-store-types.ts — public contracts
  * - candidate-geometry.ts — pure geometry hit helpers
- * - candidate-store-eager.ts — compact indexes + query methods
+ * - candidate-path-geometry.ts — path AABB / edge helpers for hit shortlists
+ * - candidate-store-indexes.ts — typed arrays, traversal, group buckets
+ * - candidate-store-spatial.ts — spatial shortlist + geometry refine
+ * - candidate-store-eager.ts — assembles indexes/spatial into CandidateStore
  * - candidate-store-lazy.ts — deferred construction shell
  */
 export { canonicalAxisToken } from "./candidate-axis-token.js";
