--- packages/core/tests/candidate-store.test.ts
@@ -1,1350 +1,0 @@
-import { fromPartial } from "@total-typescript/shoehorn";
-import { describe, expect, it, spyOn } from "bun:test";
-
-import {
-  buildCandidateStore,
-  canonicalAxisToken,
-  type CandidateDatum,
-} from "../src/candidate-store.ts";
-import type { Scene } from "../src/scene.ts";
-
-function scene(): Scene {
-  return {
-    width: 200,
-    height: 120,
-    panels: [
-      {
-        id: "panel:all",
-        x: 0,
-        y: 0,
-        width: 200,
-        height: 120,
-        strip: "",
-        axisX: [],
-        axisY: [],
-        grid: { x: [], y: [] },
-      },
-    ],
-    batches: [
-      {
-        kind: "points",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([10, 20, 10, 40, 50, 30]),
-        rowIndex: new Uint32Array([0, 1, 2]),
-        size: 3,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-      {
-        kind: "points",
-        layerIndex: 1,
-        panelIndex: 0,
-        positions: new Float32Array([10, 25, 10, 25]),
-        rowIndex: new Uint32Array([3, 4]),
-        size: 3,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-    ],
-    axes: { x: { ticks: [], title: "" }, y: { ticks: [], title: "" } },
-    grid: { x: [], y: [] },
-    legends: [],
-    theme: fromPartial<Scene["theme"]>({}),
-    title: "",
-    subtitle: "",
-    caption: "",
-  };
-}
-
-function sceneWithPoints(points: readonly (readonly [number, number])[]): Scene {
-  const result = scene();
-  result.batches = [
-    {
-      kind: "points",
-      layerIndex: 0,
-      panelIndex: 0,
-      positions: Float32Array.from(points.flat()),
-      rowIndex: Uint32Array.from(points.map((_, index) => index)),
-      size: 3,
-      alpha: 1,
-      shape: "circle",
-      fill: null,
-    },
-  ];
-  return result;
-}
-
-const data: CandidateDatum[] = [
-  { xValue: new Date(0), yValue: 20, seriesId: 0, seriesRank: 1 },
-  { xValue: new Date(0), yValue: 40, seriesId: 0, seriesRank: 1 },
-  { xValue: 50, yValue: 30, seriesId: 0, seriesRank: 1 },
-  { xValue: new Date(0), yValue: 25, seriesId: 8, seriesRank: 0 },
-  { xValue: new Date(0), yValue: 25, seriesId: 8, seriesRank: 0 },
-];
-
-describe("canonicalAxisToken", () => {
-  it("normalizes supported values and excludes invalid buckets", () => {
-    expect(canonicalAxisToken(new Date(12))).toEqual({ kind: "number", value: 12 });
-    expect(canonicalAxisToken(-0)).toEqual({ kind: "number", value: 0 });
-    expect(canonicalAxisToken("a")).toEqual({ kind: "string", value: "a" });
-    expect(canonicalAxisToken(false)).toEqual({ kind: "boolean", value: false });
-    expect(canonicalAxisToken(null)).toBeNull();
-    expect(canonicalAxisToken(Number.NaN)).toBeNull();
-    expect(canonicalAxisToken(Infinity)).toBeNull();
-  });
-});
-
-describe("candidate-store public facade", () => {
-  it("exposes only buildCandidateStore and canonicalAxisToken at runtime", async () => {
-    const facade = await import("../src/candidate-store.ts");
-    expect(Object.keys(facade).toSorted()).toEqual(
-      ["buildCandidateStore", "canonicalAxisToken"].toSorted(),
-    );
-  });
-});
-
-describe("CandidateStore", () => {
-  const store = buildCandidateStore(scene(), {
-    epoch: 7,
-    datum: ({ candidateIndex }) => data[candidateIndex]!,
-  });
-
-  it("defers semantic/index construction until interaction first reads the store", () => {
-    let resolutions = 0;
-    const lazy = buildCandidateStore(scene(), {
-      datum: ({ candidateIndex }) => {
-        resolutions++;
-        return data[candidateIndex]!;
-      },
-    });
-    expect(lazy.size).toBe(5);
-    expect(resolutions).toBe(0);
-    expect(lazy.candidate(0)?.id).toBe(0);
-    expect(resolutions).toBe(5);
-  });
-
-  it("hit-tests points through the lazy store with tolerance and panel clipping", () => {
-    let resolutions = 0;
-    const lazy = buildCandidateStore(sceneWithPoints([[0, 20]]), {
-      datum: () => {
-        resolutions++;
-        return {};
-      },
-    });
-    expect(resolutions).toBe(0);
-    expect(lazy.hitTest(5.5, 20)?.id).toBe(0);
-    expect(resolutions).toBe(1);
-    expect(lazy.hitTest(6.5, 20)).toBeNull();
-    // The rendered mark is clipped at the panel edge even though its radius
-    // extends around the point anchor.
-    expect(lazy.hitTest(-1, 20)).toBeNull();
-
-    const exactRadius = buildCandidateStore(sceneWithPoints([[10, 20]]), {
-      hitTolerance: 0,
-    });
-    expect(exactRadius.hitTest(12.5, 20)?.id).toBe(0);
-    expect(exactRadius.hitTest(13.5, 20)).toBeNull();
-  });
-
-  it("hitTest follows reverse paint order for overlapping exact geometry", () => {
-    const points = scene();
-    points.batches = [
-      {
-        kind: "points",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([10, 20]),
-        rowIndex: new Uint32Array([0]),
-        size: 5,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-      {
-        kind: "points",
-        layerIndex: 1,
-        panelIndex: 0,
-        positions: new Float32Array([14, 20]),
-        rowIndex: new Uint32Array([1]),
-        size: 5,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-    ];
-    // The later-painted point wins even though the earlier point is closer.
-    expect(buildCandidateStore(points).hitTest(10, 20)?.id).toBe(1);
-
-    const rects = scene();
-    rects.batches = [0, 1].map((layerIndex) => ({
-      kind: "rects" as const,
-      layerIndex,
-      panelIndex: 0,
-      rects: new Float32Array([10, 10, 40, 40]),
-      rowIndex: new Uint32Array([layerIndex]),
-      fill: null,
-      alpha: 1,
-    }));
-    expect(buildCandidateStore(rects).hitTest(20, 20)?.id).toBe(1);
-    expect(buildCandidateStore(rects).hitTest(5, 20)).toBeNull();
-    rects.batches = [
-      {
-        kind: "rects",
-        layerIndex: 0,
-        panelIndex: 0,
-        rects: new Float32Array([50, 50, -40, -40]),
-        rowIndex: new Uint32Array([0]),
-        fill: null,
-        alpha: 1,
-      },
-    ];
-    const negativeRectStore = buildCandidateStore(rects);
-    expect(negativeRectStore.hitTest(20, 20)?.id).toBe(0);
-    expect(negativeRectStore.queryRect(15, 15, 25, 25)).toEqual(new Uint32Array([0]));
-
-    const segments = scene();
-    segments.batches = [0, 1].map((layerIndex) => ({
-      kind: "segments" as const,
-      layerIndex,
-      panelIndex: 0,
-      segments: new Float32Array([10, 20, 80, 20]),
-      rowIndex: new Uint32Array([layerIndex]),
-      stroke: null,
-      linewidth: 2,
-      alpha: 1,
-    }));
-    expect(buildCandidateStore(segments, { hitTolerance: 2 }).hitTest(40, 22)?.id).toBe(1);
-    expect(buildCandidateStore(segments, { hitTolerance: 2 }).hitTest(40, 24)).toBeNull();
-  });
-
-  it("hitTest resolves paths by paint order and stable nearest vertex, never glyphs", () => {
-    const paths = scene();
-    paths.batches = [
-      {
-        kind: "paths",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([10, 20, 30, 20, 10, 20, 30, 20]),
-        rowIndex: new Uint32Array([0, 1, 2, 3]),
-        pathOffsets: new Uint32Array([0, 2, 4]),
-        strokes: [null, null],
-        linewidth: 2,
-        alpha: 1,
-        curve: "linear",
-      },
-    ];
-    // Later subpath wins; an equidistant edge chooses its first vertex.
-    expect(buildCandidateStore(paths).hitTest(20, 20)?.id).toBe(2);
-
-    const folded = scene();
-    folded.batches = [
-      {
-        kind: "paths",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([0, 10, 20, 10, 10, 11]),
-        rowIndex: new Uint32Array([0, 1, 2]),
-        pathOffsets: new Uint32Array([0, 3]),
-        strokes: [null],
-        linewidth: 2,
-        alpha: 1,
-        curve: "linear",
-      },
-    ];
-    // Both edges are in tolerance. Paint compatibility chooses the first edge,
-    // then its first endpoint on an equal endpoint-distance tie.
-    expect(buildCandidateStore(folded).hitTest(10, 10)?.id).toBe(0);
-
-    const area = scene();
-    area.batches = [
-      {
-        kind: "paths",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([20, 20, 80, 20, 50, 80]),
-        rowIndex: new Uint32Array([0, 1, 2]),
-        pathOffsets: new Uint32Array([0, 3]),
-        strokes: [null],
-        fills: [null],
-        closed: true,
-        linewidth: 0,
-        alpha: 1,
-        curve: "linear",
-      },
-    ];
-    const areaStore = buildCandidateStore(area);
-    expect(areaStore.hitTest(25, 25)?.id).toBe(0);
-    // Filled areas have no visible stroke; pointer tolerance must not create
-    // an invisible interactive band outside the polygon.
-    expect(areaStore.hitTest(50, 18)).toBeNull();
-
-    const glyphs = scene();
-    glyphs.batches = [
-      {
-        kind: "glyphs",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([20, 20]),
-        rowIndex: new Uint32Array([0]),
-        text: ["label"],
-        fill: null,
-        size: 12,
-        alpha: 1,
-      },
-    ];
-    expect(buildCandidateStore(glyphs).hitTest(20, 20)).toBeNull();
-  });
-
-  it("chooses the nearest semantic anchor on a shared tessellated edge", () => {
-    const tessellated = scene();
-    tessellated.batches = [
-      {
-        kind: "paths",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([0, 10, 10, 10, 20, 10, 30, 10]),
-        rowIndex: new Uint32Array([0, 1, 1, 1]),
-        semanticAnchors: new Uint8Array([1, 0, 0, 1]),
-        semanticIndex: new Uint32Array([0, 1, 1, 1]),
-        pathOffsets: new Uint32Array([0, 4]),
-        strokes: [null],
-        linewidth: 2,
-        alpha: 1,
-        curve: "linear",
-      },
-    ];
-    const tessellatedStore = buildCandidateStore(tessellated, {
-      datum: ({ primitiveIndex }) => ({ xValue: primitiveIndex }),
-    });
-
-    expect(tessellatedStore.hitTest(2, 10)).toMatchObject({ rowIndex: 0, primitiveIndex: 0 });
-    expect(tessellatedStore.hitTest(28, 10)).toMatchObject({ rowIndex: 1, primitiveIndex: 3 });
-  });
-
-  it("hitTest preserves exact containment when semantic nearest is too far", () => {
-    const largeRect = scene();
-    largeRect.batches = [
-      {
-        kind: "rects",
-        layerIndex: 0,
-        panelIndex: 0,
-        rects: new Float32Array([10, 10, 160, 90]),
-        rowIndex: new Uint32Array([0]),
-        fill: null,
-        alpha: 1,
-      },
-    ];
-    const candidates = buildCandidateStore(largeRect);
-    expect(candidates.nearest(20, 90, { mode: "xy", maxDistance: 5 })).toBeNull();
-    expect(candidates.hitTest(20, 90)?.id).toBe(0);
-  });
-
-  it("uses exact geometry for containment and rectangle intersections", () => {
-    const rectScene = scene();
-    rectScene.batches = [
-      {
-        kind: "rects",
-        layerIndex: 0,
-        panelIndex: 0,
-        rects: new Float32Array([20, 20, 40, 40]),
-        rowIndex: new Uint32Array([0]),
-        fill: null,
-        alpha: 1,
-      },
-    ];
-    const rectStore = buildCandidateStore(rectScene, { datum: () => ({ xValue: 1, yValue: 2 }) });
-    expect(rectStore.nearest(22, 55, { mode: "exact", maxDistance: 0 })?.id).toBe(0);
-    expect(rectStore.queryRect(15, 30, 25, 35)).toEqual(new Uint32Array([0]));
-    expect(rectStore.candidate(0)?.autoMode).toBe("exact");
-    expect(rectStore.nearest(22, 55, { mode: "auto", maxDistance: 0 })?.mode).toBe("exact");
-  });
-
-  it("infers a semantic dominant axis for standalone segments", () => {
-    const segmentScene = scene();
-    segmentScene.batches = [
-      {
-        kind: "segments",
-        layerIndex: 0,
-        panelIndex: 0,
-        segments: new Float32Array([20, 0, 20, 100, 0, 40, 100, 40]),
-        rowIndex: new Uint32Array([0, 1]),
-        stroke: null,
-        linewidth: 1,
-        alpha: 1,
-      },
-    ];
-    const segmentStore = buildCandidateStore(segmentScene, {
-      datum: ({ primitiveIndex }) => ({
-        xValue: primitiveIndex === 0 ? 20 : null,
-        yValue: primitiveIndex === 1 ? 40 : null,
-      }),
-    });
-
-    expect(segmentStore.candidate(0)?.autoMode).toBe("x");
-    expect(segmentStore.candidate(1)?.autoMode).toBe("y");
-    expect(segmentStore.nearest(22, 90, { mode: "auto", maxDistance: 3 })).toMatchObject({
-      id: 0,
-      mode: "x",
-    });
-    expect(segmentStore.nearest(90, 42, { mode: "auto", maxDistance: 3 })).toMatchObject({
-      id: 1,
-      mode: "y",
-    });
-  });
-
-  it("uses polygon containment for exact filled-path lookup", () => {
-    const areaScene = scene();
-    areaScene.batches = [
-      {
-        kind: "paths",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([20, 20, 80, 20, 50, 80]),
-        rowIndex: new Uint32Array([0, 1, 2]),
-        pathOffsets: new Uint32Array([0, 3]),
-        strokes: [null],
-        fills: [null],
-        closed: true,
-        linewidth: 0,
-        alpha: 1,
-        curve: "linear",
-      },
-    ];
-    const areaStore = buildCandidateStore(areaScene);
-    expect(areaStore.nearest(25, 25, { mode: "exact", maxDistance: 0 })?.id).toBe(0);
-  });
-
-  it("does not form stroke segments across multi-subpath boundaries", () => {
-    // Two disjoint horizontal strokes. The last vertex of path 0 is (20,0);
-    // the first of path 1 is (0,50). A false edge between them would pass near (10,25).
-    const multi = scene();
-    multi.batches = [
-      {
-        kind: "paths",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([0, 0, 10, 0, 20, 0, 0, 50, 10, 50, 20, 50]),
-        rowIndex: new Uint32Array([0, 1, 2, 3, 4, 5]),
-        pathOffsets: new Uint32Array([0, 3, 6]),
-        strokes: [null, null],
-        linewidth: 2,
-        alpha: 1,
-        curve: "linear",
-      },
-    ];
-    const multiStore = buildCandidateStore(multi, {
-      datum: ({ primitiveIndex }) => ({
-        xValue: primitiveIndex,
-        yValue: primitiveIndex < 3 ? 0 : 50,
-        // Force exact geometry so pathRange/samePath are on the refine path.
-        autoMode: "exact",
-      }),
-    });
-    // Midpoint of the false cross-boundary edge must not hit.
-    expect(multiStore.nearest(10, 25, { mode: "exact", maxDistance: 3 })).toBeNull();
-    // Real stroke on path 0 (probe near the middle vertex of the first subpath).
-    const onPath0 = multiStore.nearest(10, 1, { mode: "exact", maxDistance: 3 });
-    expect(onPath0).not.toBeNull();
-    expect(onPath0!.id).toBeLessThan(3);
-    // Real stroke on path 1.
-    const onPath1 = multiStore.nearest(10, 51, { mode: "exact", maxDistance: 3 });
-    expect(onPath1).not.toBeNull();
-    expect(onPath1!.id).toBeGreaterThanOrEqual(3);
-    // Brush covering only the false edge should be empty; covering a real edge is not.
-    expect(multiStore.queryRect(8, 20, 12, 30)).toEqual(new Uint32Array());
-    const brushed = multiStore.queryRect(8, -2, 12, 2);
-    expect([...brushed].every((id) => id < 3)).toBe(true);
-    expect(brushed.length).toBeGreaterThan(0);
-  });
-
-  it("rejects segment bounding boxes that do not geometrically intersect a brush", () => {
-    const segmentScene = scene();
-    segmentScene.batches = [
-      {
-        kind: "segments",
-        layerIndex: 0,
-        panelIndex: 0,
-        segments: new Float32Array([0, 0, 10, 10]),
-        rowIndex: new Uint32Array([0]),
-        stroke: null,
-        linewidth: 1,
-        alpha: 1,
-      },
-    ];
-    const segmentStore = buildCandidateStore(segmentScene);
-    expect(segmentStore.queryRect(0, 9, 1, 10)).toEqual(new Uint32Array());
-    expect(segmentStore.queryRect(4, 4, 6, 6)).toEqual(new Uint32Array([0]));
-  });
-
-  it("releases lazy resolvers and initialized candidate arrays on dispose", () => {
-    let resolutions = 0;
-    const lazy = buildCandidateStore(scene(), {
-      datum: () => {
-        resolutions++;
-        return {};
-      },
-    });
-    lazy.dispose();
-    expect(lazy.size).toBe(0);
-    expect(lazy.candidate(0)).toBeNull();
-    expect(lazy.x).toHaveLength(0);
-    expect(resolutions).toBe(0);
-
-    const initialized = buildCandidateStore(scene());
-    expect(initialized.candidate(0)).not.toBeNull();
-    initialized.dispose();
-    expect(initialized.size).toBe(0);
-    expect(initialized.candidate(0)).toBeNull();
-    expect(initialized.x).toHaveLength(0);
-  });
-
-  it("owns typed anchors and exposes stable candidate facts", () => {
-    expect(store.x).toBeInstanceOf(Float32Array);
-    expect(store.y).toBeInstanceOf(Float32Array);
-    expect(store.size).toBe(5);
-    expect(store.candidate(0)).toMatchObject({
-      id: 0,
-      epoch: 7,
-      panelId: "panel:all",
-      rowIndex: 0,
-      seriesId: 0,
-    });
-  });
-
-  it("finds dominant-axis and euclidean nearest candidates", () => {
-    expect(store.nearest(12, 39, { mode: "x", maxDistance: 3 })?.id).toBe(1);
-    expect(store.nearest(48, 31, { mode: "xy", maxDistance: 4 })?.id).toBe(2);
-    expect(store.nearest(80, 80, { mode: "xy", maxDistance: 2 })).toBeNull();
-  });
-
-  it("excludes invalid logical values from dominant-axis inspection", () => {
-    const invalid = [
-      ...data.slice(0, 3),
-      { xValue: Number.NaN, yValue: 2, seriesId: 9 },
-      ...data.slice(3),
-    ];
-    const withInvalid = scene();
-    const points = withInvalid.batches[0]!;
-    if (points.kind !== "points") throw new Error("fixture");
-    points.positions = new Float32Array([...points.positions, 80, 2]);
-    points.rowIndex = new Uint32Array([...points.rowIndex, 5]);
-    const invalidStore = buildCandidateStore(withInvalid, {
-      datum: ({ candidateIndex }) => invalid[candidateIndex]!,
-    });
-    expect(invalidStore.nearest(80, 2, { mode: "x", maxDistance: 1 })).toBeNull();
-  });
-
-  it("returns a compact canonical bucket and one representative per series", () => {
-    const grouped = store.group(1, "x");
-    expect(grouped?.axisValue).toEqual(new Date(0));
-    expect(grouped?.memberIds).toEqual(new Uint32Array([3, 1]));
-    expect(grouped?.focusId).toBe(1);
-    expect(grouped?.range.end).toBeGreaterThan(grouped?.range.start ?? 0);
-  });
-
-  it("cycles coincident marks, traverses deterministically, and returns integer rect ids", () => {
-    // Coincident pair at (10, 25): paint/source order is ids 3 then 4.
-    expect(store.cycle(3, 1)).toBe(4);
-    expect(store.cycle(4, 1)).toBe(3);
-    expect(store.cycle(3, -1)).toBe(4);
-    expect(store.cycle(4, -1)).toBe(3);
-    expect(store.cycle(3, 2)).toBe(3);
-    expect(store.cycle(3, 0)).toBe(3);
-    // Singleton stack: step is a no-op (no retained one-element stack required).
-    expect(store.cycle(0, 1)).toBe(0);
-    expect(store.cycle(0, -3)).toBe(0);
-    // Invalid seed returns null (bounds + non-integer, matching other store entry points).
-    expect(store.cycle(-1, 1)).toBeNull();
-    expect(store.cycle(5, 1)).toBeNull();
-    expect(store.cycle(1.5, 1)).toBeNull();
-    expect(store.cycle(Number.NaN, 1)).toBeNull();
-    // Non-finite / non-integral step falls back to the seed (prior contract).
-    expect(store.cycle(3, Number.NaN)).toBe(3);
-    expect(store.cycle(3, Number.POSITIVE_INFINITY)).toBe(3);
-    expect(store.cycle(3, 1.5)).toBe(3);
-    expect(store.traverse(0, "down")).toBe(4);
-    expect(store.queryRect(5, 15, 15, 45)).toEqual(new Uint32Array([0, 3, 4, 1]));
-  });
-
-  // Traversal order for this fixture (panel → y → x → batch → primitive):
-  // [0, 3, 4, 2, 1]
-  it("walks next/previous with wrap-around and falls back for unknown start ids", () => {
-    expect(store.traverse(null, "next")).toBe(0);
-    expect(store.traverse(null, "previous")).toBe(0);
-    expect(store.traverse(0, "first")).toBe(0);
-    expect(store.traverse(0, "last")).toBe(1);
-    expect(store.traverse(0, "next")).toBe(3);
-    expect(store.traverse(3, "next")).toBe(4);
-    expect(store.traverse(4, "next")).toBe(2);
-    expect(store.traverse(2, "next")).toBe(1);
-    expect(store.traverse(1, "next")).toBe(0);
-    expect(store.traverse(0, "previous")).toBe(1);
-    expect(store.traverse(1, "previous")).toBe(2);
-    expect(store.traverse(2, "previous")).toBe(4);
-    // CandidateStore owns modular keyboard jumps; callers do not materialize
-    // traversal order to calculate an index themselves.
-    expect(store.traverse(null, "next", 2)).toBe(3);
-    expect(store.traverse(0, "next", 3)).toBe(2);
-    expect(store.traverse(0, "previous", 2)).toBe(2);
-    expect(store.traverse(999, "next")).toBe(0);
-    expect(store.traverse(-1, "previous")).toBe(0);
-  });
-
-  it("keeps spatial directions independent of sequential rank", () => {
-    // Spatial uses geometry, not traversal order.
-    expect(store.traverse(0, "down")).toBe(4);
-    expect(store.traverse(0, "right")).toBe(2);
-    // From (50,30), nearest left is the coincident pair at x=10; topmost id 4 wins.
-    expect(store.traverse(2, "left")).toBe(4);
-    expect(store.traverse(1, "up")).toBe(2);
-  });
-});
-
-describe("candidate cycle hot path", () => {
-  it("uses prebuilt coincident stacks without scanning or indexOf on each step", () => {
-    // Three-way stack at (10, 10) plus a singleton — build order is paint/source (id asc).
-    const plotScene = sceneWithPoints([
-      [10, 10],
-      [20, 20],
-      [10, 10],
-      [10, 10],
-    ]);
-    const hot = buildCandidateStore(plotScene, {
-      datum: ({ primitiveIndex }) => ({
-        xValue: primitiveIndex,
-        yValue: primitiveIndex,
-      }),
-    });
-    // Force the lazy construction boundary before policing resolution.
-    void hot.x;
-    const arrayIndexOfSpy = spyOn(Array.prototype, "indexOf").mockImplementation(function (
-      this: unknown[],
-      ...args: Parameters<Array<unknown>["indexOf"]>
-    ) {
-      throw new Error(`cycle used Array.indexOf(${String(args[0])})`);
-    });
-    const typedIndexOfSpy = spyOn(Uint32Array.prototype, "indexOf").mockImplementation(function (
-      this: Uint32Array,
-      ...args: Parameters<Uint32Array["indexOf"]>
-    ) {
-      throw new Error(`cycle used Uint32Array.indexOf(${String(args[0])})`);
-    });
-    try {
-      expect(hot.cycle(0, 1)).toBe(2);
-      expect(hot.cycle(2, 1)).toBe(3);
-      expect(hot.cycle(3, 1)).toBe(0);
-      expect(hot.cycle(0, -1)).toBe(3);
-      expect(hot.cycle(1, 5)).toBe(1);
-      expect(hot.cycle(-1)).toBeNull();
-      expect(hot.cycle(99)).toBeNull();
-    } finally {
-      arrayIndexOfSpy.mockRestore();
-      typedIndexOfSpy.mockRestore();
-    }
-  });
-});
-
-describe("candidate traversal hot path", () => {
-  it("uses O(1) rank lookup for next/previous without scanning traversal", () => {
-    const plotScene = sceneWithPoints([
-      [10, 10],
-      [20, 20],
-      [30, 30],
-      [40, 40],
-    ]);
-    const hot = buildCandidateStore(plotScene, {
-      datum: ({ primitiveIndex }) => ({
-        xValue: primitiveIndex,
-        yValue: primitiveIndex,
-      }),
-    });
-    // Force the lazy construction boundary before policing resolution.
-    void hot.x;
-    const indexOfSpy = spyOn(Uint32Array.prototype, "indexOf").mockImplementation(function (
-      this: Uint32Array,
-      ...args: Parameters<Uint32Array["indexOf"]>
-    ) {
-      throw new Error(`traverse used linear indexOf(${String(args[0])})`);
-    });
-    try {
-      expect(hot.traverse(0, "next")).toBe(1);
-      expect(hot.traverse(1, "previous")).toBe(0);
-      expect(hot.traverse(3, "next")).toBe(0);
-      expect(hot.traverse(0, "previous")).toBe(3);
-      expect(hot.traverse(null)).toBe(0);
-      expect(hot.traverse(0, "first")).toBe(0);
-      expect(hot.traverse(0, "last")).toBe(3);
-      // Spatial still allowed to scan candidates by id; only sequential rank is constrained.
-      expect(hot.traverse(0, "down")).toBe(1);
-    } finally {
-      indexOfSpy.mockRestore();
-    }
-  });
-
-  it("directional traverse picks nearest primary run without scanning all candidates", () => {
-    // Horizontal line of 2048 points; seed in the middle. Linear scan would
-    // touch every id; binary search only inspects O(log n + k) for k=1.
-    const n = 2048;
-    const points: (readonly [number, number])[] = [];
-    for (let i = 0; i < n; i++) points.push([i, 0]);
-    const hot = buildCandidateStore(sceneWithPoints(points), {
-      datum: ({ primitiveIndex }) => ({
-        xValue: primitiveIndex,
-        yValue: 0,
-      }),
-    });
-    void hot.x;
-    const seed = 1000;
-    expect(hot.traverse(seed, "right")).toBe(seed + 1);
-    expect(hot.traverse(seed, "left")).toBe(seed - 1);
-    expect(hot.traverse(0, "left")).toBe(0); // nothing further left
-    expect(hot.traverse(n - 1, "right")).toBe(n - 1);
-  });
-
-  it("directional traverse prefers topmost paint order when primary and orth both tie", () => {
-    // Three points at x=20 with y=0,10,0; seed at (0,0). Right nearest primary
-    // is x=20; among y=0 ties (orth 0), the later-painted higher id wins.
-    const hot = buildCandidateStore(
-      sceneWithPoints([
-        [0, 0],
-        [20, 0],
-        [20, 10],
-        [20, 0],
-      ]),
-      {
-        datum: ({ primitiveIndex }) => ({
-          xValue: primitiveIndex,
-          yValue: primitiveIndex,
-        }),
-      },
-    );
-    void hot.x;
-    expect(hot.traverse(0, "right")).toBe(3);
-  });
-});
-
-describe("candidate grouping hot path", () => {
-  it("uses preordered series boundaries without sorting during resolution", () => {
-    const plotScene = sceneWithPoints([
-      [10, 30],
-      [10, 10],
-      [10, 20],
-    ]);
-    const store = buildCandidateStore(plotScene, {
-      datum: ({ primitiveIndex }) => ({
-        xValue: 1,
-        yValue: primitiveIndex,
-        seriesId: primitiveIndex,
-        seriesRank: 2 - primitiveIndex,
-      }),
-    });
-    // Force the lazy construction/sort boundary before policing resolution.
-    void store.x;
-    const sortSpy = spyOn(Array.prototype, "toSorted").mockImplementation(function () {
-      throw new Error("group resolution sorted");
-    });
-    try {
-      expect([...store.group(0, "x")!.memberIds]).toEqual([2, 1, 0]);
-    } finally {
-      sortSpy.mockRestore();
-    }
-  });
-
-  it("does not produce an axis target for an invalid logical bucket", () => {
-    const plotScene = sceneWithPoints([[10, 10]]);
-    const store = buildCandidateStore(plotScene, {
-      datum: () => ({ xValue: Number.NaN, yValue: 1 }),
-    });
-    expect(store.group(0, "x")).toBeNull();
-    expect(store.nearest(10, 10, { mode: "x", maxDistance: 100 })).toBeNull();
-  });
-
-  it("keeps the seed as its own series representative", () => {
-    const plotScene = sceneWithPoints([
-      [10, 10],
-      [10, 50],
-      [10, 30],
-    ]);
-    // Two series at the same x: seed series 0 has two points; series 1 has one.
-    const store = buildCandidateStore(plotScene, {
-      datum: ({ primitiveIndex }) => ({
-        xValue: "a",
-        yValue: primitiveIndex * 10,
-        seriesId: primitiveIndex === 1 ? 1 : 0,
-        seriesRank: primitiveIndex === 1 ? 1 : 0,
-      }),
-    });
-    const grouped = store.group(2, "x");
-    expect(grouped).not.toBeNull();
-    // Seed id 2 is series 0; that series' member must stay the seed (not the
-    // orth-closest sibling id 0 at y=10).
-    expect(grouped!.focusId).toBe(2);
-    expect([...grouped!.memberIds]).toContain(2);
-  });
-
-  it("resolves a large non-seed series with O(log M) orth probes", () => {
-    // Seed series is a singleton; other series has M points at the same x
-    // with increasing y. group(seed) must binary-search the large series —
-    // a linear scan would call Math.abs ~2*(M-1) times on that series alone.
-    const M = 2048;
-    const points: (readonly [number, number])[] = [[0, 0]]; // seed series
-    for (let i = 0; i < M; i++) points.push([0, i]); // other series
-    const plotScene = sceneWithPoints(points);
-    const store = buildCandidateStore(plotScene, {
-      datum: ({ primitiveIndex }) => ({
-        xValue: 1,
-        yValue: primitiveIndex,
-        seriesId: primitiveIndex === 0 ? 0 : 1,
-        seriesRank: primitiveIndex === 0 ? 0 : 1,
-      }),
-    });
-    void store.x;
-    const absSpy = spyOn(Math, "abs");
-    try {
-      const grouped = store.group(0, "x");
-      expect(grouped).not.toBeNull();
-      // Closest in the large series to seed y=0 is the first member (y=0).
-      expect([...grouped!.memberIds]).toContain(1);
-      expect(absSpy.mock.calls.length).toBeLessThan(80);
-    } finally {
-      absSpy.mockRestore();
-    }
-  });
-});
-
-describe("candidate spatial query hot path (issue #214)", () => {
-  function largePointScene(count: number): Scene {
-    const cols = Math.ceil(Math.sqrt(count));
-    const points: (readonly [number, number])[] = [];
-    for (let i = 0; i < count; i++) {
-      points.push([(i % cols) * 10, Math.floor(i / cols) * 10]);
-    }
-    return sceneWithPoints(points);
-  }
-
-  it("xy nearest does not Math.hypot every candidate on a large point cloud", () => {
-    // Complexity guard: linear nearest is O(n) hypot; spatial shortlist is O(log n + k).
-    const count = 8_000;
-    const store = buildCandidateStore(largePointScene(count), {
-      datum: ({ primitiveIndex }) => ({
-        xValue: primitiveIndex,
-        yValue: primitiveIndex,
-      }),
-    });
-    void store.x;
-
-    const originalHypot = Math.hypot;
-    let hypotCalls = 0;
-    const hypotSpy = spyOn(Math, "hypot").mockImplementation(function (
-      this: void,
-      ...args: Parameters<typeof Math.hypot>
-    ) {
-      hypotCalls++;
-      return originalHypot(...args);
-    });
-    try {
-      // Target near (10,10) — grid id 1 at (10,0) or id ~cols at (0,10); nearest is id 1 or nearby.
-      const hit = store.nearest(10, 10, { mode: "xy", maxDistance: 6 });
-      expect(hit).not.toBeNull();
-      expect(hit!.distance).toBeLessThanOrEqual(6);
-      // Shortlist must be far smaller than a full scan (budget leaves room for refine + ties).
-      expect(hypotCalls).toBeLessThan(200);
-      expect(hypotCalls).toBeLessThan(count / 10);
-    } finally {
-      hypotSpy.mockRestore();
-    }
-  });
-
-  it("hitTest does not expand a dense point query by one oversized point", () => {
-    const count = 10_000;
-    const dense = new Float32Array(count * 2);
-    for (let i = 0; i < count; i++) {
-      dense[i * 2] = i % 100;
-      dense[i * 2 + 1] = Math.floor(i / 100);
-    }
-    const plotScene = scene();
-    plotScene.batches = [
-      {
-        kind: "points",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: dense,
-        rowIndex: Uint32Array.from({ length: count }, (_, index) => index),
-        size: 1,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-      {
-        kind: "points",
-        layerIndex: 1,
-        panelIndex: 0,
-        positions: new Float32Array([190, 100]),
-        rowIndex: new Uint32Array([count]),
-        size: 1_000,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-    ];
-    const store = buildCandidateStore(plotScene);
-    void store.x;
-    const hypot = spyOn(Math, "hypot");
-    hypot.mockClear();
-    expect(store.hitTest(10, 10)?.id).toBe(count);
-    expect(hypot).toHaveBeenCalledTimes(1);
-    hypot.mockRestore();
-  });
-
-  it("queryRect does not read every batch for a tight brush on a large point cloud", () => {
-    const count = 8_000;
-    const cols = Math.ceil(Math.sqrt(count));
-    const positions = new Float32Array(count * 2);
-    for (let i = 0; i < count; i++) {
-      positions[i * 2] = (i % cols) * 10;
-      positions[i * 2 + 1] = Math.floor(i / cols) * 10;
-    }
-    const rawBatches: Scene["batches"] = [
-      {
-        kind: "points",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions,
-        rowIndex: Uint32Array.from({ length: count }, (_, index) => index),
-        size: 3,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-    ];
-    let batchIndexReads = 0;
-    const batches = new Proxy(rawBatches, {
-      get(target, property, receiver): unknown {
-        if (typeof property === "string" && /^\d+$/.test(property)) batchIndexReads++;
-        return Reflect.get(target, property, receiver) as unknown;
-      },
-    });
-    const plotScene = scene();
-    plotScene.batches = batches;
-    const store = buildCandidateStore(plotScene);
-    // Construction may touch batches; only police the query.
-    void store.x;
-    batchIndexReads = 0;
-
-    const hits = store.queryRect(5, 5, 25, 25);
-    // Grid points in [5,25]² with 10px spacing: (10,10), (20,10), (10,20), (20,20) → 4 hits.
-    expect([...hits].toSorted((a, b) => a - b)).toEqual([
-      cols + 1,
-      cols + 2,
-      2 * cols + 1,
-      2 * cols + 2,
-    ]);
-    // Linear queryRect consulted scene.batches once per candidate; shortlist stays O(k).
-    expect(batchIndexReads).toBeLessThan(200);
-    expect(batchIndexReads).toBeLessThan(count / 10);
-  });
-
-  it("preserves topmost tie-break and panel filter with spatial shortlist", () => {
-    const multi = scene();
-    multi.panels = [
-      {
-        id: "a",
-        x: 0,
-        y: 0,
-        width: 100,
-        height: 100,
-        strip: "",
-        axisX: [],
-        axisY: [],
-        grid: { x: [], y: [] },
-      },
-      {
-        id: "b",
-        x: 200,
-        y: 0,
-        width: 100,
-        height: 100,
-        strip: "",
-        axisX: [],
-        axisY: [],
-        grid: { x: [], y: [] },
-      },
-    ];
-    multi.batches = [
-      {
-        kind: "points",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([10, 10, 10, 10]),
-        rowIndex: new Uint32Array([0, 1]),
-        size: 3,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-      {
-        kind: "points",
-        layerIndex: 1,
-        panelIndex: 1,
-        positions: new Float32Array([10, 10]),
-        rowIndex: new Uint32Array([2]),
-        size: 3,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-    ];
-    const store = buildCandidateStore(multi);
-    // Coincident panel-a points: reverse paint order → higher id wins equal distance.
-    expect(store.nearest(10, 10, { mode: "xy", maxDistance: 1 })?.id).toBe(1);
-    expect(store.nearest(10, 10, { mode: "xy", maxDistance: 1, panelId: "a" })?.id).toBe(1);
-    expect(store.nearest(210, 10, { mode: "xy", maxDistance: 1, panelId: "b" })?.id).toBe(2);
-    expect(store.nearest(10, 10, { mode: "xy", maxDistance: 1, panelId: "b" })).toBeNull();
-    expect([...store.queryRect(5, 5, 15, 15, "a")].toSorted((a, b) => a - b)).toEqual([0, 1]);
-    expect([...store.queryRect(205, 5, 215, 15, "b")]).toEqual([2]);
-  });
-
-  it("auto mode still finds dominant-axis point hits far from the probe orthogonally", () => {
-    // Boxplot outliers and similar points can inherit autoMode "x": match on
-    // dominant-axis distance only. Spatial shortlist must include x/y strips,
-    // not only a euclidean square around maxPointReach.
-    const plotScene = sceneWithPoints([
-      [50, 10],
-      [50, 200],
-    ]);
-    const store = buildCandidateStore(plotScene, {
-      datum: ({ primitiveIndex }) => ({
-        xValue: 50,
-        yValue: primitiveIndex === 0 ? 10 : 200,
-        autoMode: "x",
-      }),
-    });
-    // Probe at same x as both points, but 150px from the nearer in y — outside
-    // size+3 euclidean reach, yet within x maxDistance of 5.
-    expect(store.nearest(52, 100, { mode: "auto", maxDistance: 5 })).toMatchObject({
-      id: 0,
-      mode: "x",
-      distance: 2,
-    });
-    expect(store.nearest(52, 100, { mode: "x", maxDistance: 5 })?.id).toBe(0);
-  });
-
-  it("exact nearest still hits large rects whose anchors are far from the probe", () => {
-    const rectScene = scene();
-    rectScene.batches = [
-      {
-        kind: "rects",
-        layerIndex: 0,
-        panelIndex: 0,
-        // Wide/tall bar: anchor is center-x / top-y (100, 0).
-        rects: new Float32Array([0, 0, 200, 100]),
-        rowIndex: new Uint32Array([0]),
-        fill: null,
-        alpha: 1,
-      },
-      {
-        kind: "points",
-        layerIndex: 1,
-        panelIndex: 0,
-        // Far from the rect interior probe used below.
-        positions: new Float32Array([5, 5]),
-        rowIndex: new Uint32Array([1]),
-        size: 2,
-        alpha: 1,
-        shape: "circle",
-        fill: null,
-      },
-    ];
-    const store = buildCandidateStore(rectScene, {
-      datum: ({ kind }) => ({ xValue: 1, yValue: 2, autoMode: kind === "rects" ? "exact" : "xy" }),
-    });
-    // Probe inside the bar, far from its top-center anchor — spatial shortlist by
-    // anchor radius alone would miss this unless extended geometry is refined.
-    expect(store.nearest(150, 80, { mode: "exact", maxDistance: 0 })?.id).toBe(0);
-    expect(store.queryRect(140, 70, 160, 90)).toEqual(new Uint32Array([0]));
-    expect(store.queryRect(0, 0, 10, 10)).toEqual(new Uint32Array([0, 1]));
-  });
-
-  it("exact nearest does not refine every far rect on a large bar field", () => {
-    // Complexity guard: extended geometry used to force-add all E rects into the
-    // shortlist. AABB spatial shortlist must keep refine O(log E + k).
-    const count = 4_000;
-    const rects = new Float32Array(count * 4);
-    for (let i = 0; i < count; i++) {
-      // 2×2 bars spaced on a grid far from the probe at (5,5).
-      const col = i % 50;
-      const row = Math.floor(i / 50);
-      rects[i * 4] = 100 + col * 10;
-      rects[i * 4 + 1] = 100 + row * 10;
-      rects[i * 4 + 2] = 2;
-      rects[i * 4 + 3] = 2;
-    }
-    // One bar covering the probe, anchor far from (10,10).
-    rects[0] = 0;
-    rects[1] = 0;
-    rects[2] = 20;
-    rects[3] = 20;
-    const plotScene = scene();
-    plotScene.batches = [
-      {
-        kind: "rects",
-        layerIndex: 0,
-        panelIndex: 0,
-        rects,
-        rowIndex: Uint32Array.from({ length: count }, (_, index) => index),
-        fill: null,
-        alpha: 1,
-      },
-    ];
-    let batchIndexReads = 0;
-    const batches = new Proxy(plotScene.batches, {
-      get(target, property, receiver): unknown {
-        if (typeof property === "string" && /^\d+$/.test(property)) batchIndexReads++;
-        return Reflect.get(target, property, receiver) as unknown;
-      },
-    });
-    plotScene.batches = batches;
-    const store = buildCandidateStore(plotScene, {
-      datum: () => ({ xValue: 1, yValue: 2, autoMode: "exact" }),
-    });
-    void store.x;
-    batchIndexReads = 0;
-
-    const hit = store.nearest(10, 10, { mode: "exact", maxDistance: 0 });
-    expect(hit?.id).toBe(0);
-    // Linear extended scan would touch batches once per rect (~count).
-    expect(batchIndexReads).toBeLessThan(200);
-    expect(batchIndexReads).toBeLessThan(count / 10);
-
-    batchIndexReads = 0;
-    expect(store.hitTest(10, 10)?.id).toBe(0);
-    expect(batchIndexReads).toBeLessThan(200);
-    expect(batchIndexReads).toBeLessThan(count / 10);
-  });
-
-  it("hitTest does not refine every far segment on a dense field", () => {
-    const count = 4_000;
-    const segments = new Float32Array(count * 4);
-    for (let i = 0; i < count; i++) {
-      const x = 100 + (i % 50) * 10;
-      const y = 100 + Math.floor(i / 50) * 10;
-      segments.set([x, y, x + 2, y], i * 4);
-    }
-    segments.set([0, 10, 20, 10], 0);
-    const plotScene = scene();
-    plotScene.batches = [
-      {
-        kind: "segments",
-        layerIndex: 0,
-        panelIndex: 0,
-        segments,
-        rowIndex: Uint32Array.from({ length: count }, (_, index) => index),
-        stroke: null,
-        linewidth: 1,
-        alpha: 1,
-      },
-    ];
-    let batchIndexReads = 0;
-    plotScene.batches = new Proxy(plotScene.batches, {
-      get(target, property, receiver): unknown {
-        if (typeof property === "string" && /^\d+$/.test(property)) batchIndexReads++;
-        return Reflect.get(target, property, receiver) as unknown;
-      },
-    });
-    const store = buildCandidateStore(plotScene);
-    void store.x;
-    batchIndexReads = 0;
-    expect(store.hitTest(10, 10)?.id).toBe(0);
-    expect(batchIndexReads).toBeLessThan(200);
-    expect(batchIndexReads).toBeLessThan(count / 10);
-  });
-
-  it("one giant rect does not force-refine every small far rect", () => {
-    // Size-classed AABB trees: a full-plot bar must not expand the small-bar class.
-    const count = 2_000;
-    const rects = new Float32Array((count + 1) * 4);
-    // Giant bar covering most of the panel (own size class).
-    rects[0] = 0;
-    rects[1] = 0;
-    rects[2] = 200;
-    rects[3] = 120;
-    for (let i = 0; i < count; i++) {
-      const col = i % 40;
-      const row = Math.floor(i / 40);
-      rects[(i + 1) * 4] = 300 + col * 8;
-      rects[(i + 1) * 4 + 1] = 300 + row * 8;
-      rects[(i + 1) * 4 + 2] = 2;
-      rects[(i + 1) * 4 + 3] = 2;
-    }
-    const plotScene = scene();
-    plotScene.batches = [
-      {
-        kind: "rects",
-        layerIndex: 0,
-        panelIndex: 0,
-        rects,
-        rowIndex: Uint32Array.from({ length: count + 1 }, (_, index) => index),
-        fill: null,
-        alpha: 1,
-      },
-    ];
-    let batchIndexReads = 0;
-    plotScene.batches = new Proxy(plotScene.batches, {
-      get(target, property, receiver): unknown {
-        if (typeof property === "string" && /^\d+$/.test(property)) batchIndexReads++;
-        return Reflect.get(target, property, receiver) as unknown;
-      },
-    });
-    const store = buildCandidateStore(plotScene, {
-      datum: () => ({ xValue: 1, yValue: 2, autoMode: "exact" }),
-    });
-    void store.x;
-    batchIndexReads = 0;
-    // Probe on the giant bar; far small rects must stay out of refine.
-    expect(store.nearest(100, 60, { mode: "exact", maxDistance: 0 })?.id).toBe(0);
-    expect(batchIndexReads).toBeLessThan(50);
-    expect(batchIndexReads).toBeLessThan(count / 20);
-  });
-
-  it("exact nearest hits a stroked path mid-edge via incident-edge AABB", () => {
-    // Long horizontal polyline: probe near the middle of an interior edge.
-    const n = 20;
-    const positions = new Float32Array(n * 2);
-    for (let i = 0; i < n; i++) {
-      positions[i * 2] = i * 10;
-      positions[i * 2 + 1] = 50;
-    }
-    const plotScene = scene();
-    plotScene.batches = [
-      {
-        kind: "paths",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions,
-        rowIndex: Uint32Array.from({ length: n }, (_, index) => index),
-        pathOffsets: new Uint32Array([0, n]),
-        strokes: [null],
-        linewidth: 2,
-        alpha: 1,
-        curve: "linear",
-      },
-    ];
-    const store = buildCandidateStore(plotScene, {
-      datum: () => ({ xValue: 1, yValue: 2, autoMode: "exact" }),
-    });
-    // Midpoint of edge between vertices 5 (50,50) and 6 (60,50).
-    const hit = store.nearest(55, 51, { mode: "exact", maxDistance: 3 });
-    expect(hit).not.toBeNull();
-    expect(hit!.id === 5 || hit!.id === 6).toBe(true);
-  });
-
-  it("exact nearest does not refine every vertex of a far dense stroked polyline", () => {
-    // Complexity guard: full-subpath AABB tagged every vertex of a long series
-    // into one giant size class → Θ(V) refine. Incident-edge AABB keeps shortlist
-    // O(log V + k) for a distant probe.
-    const count = 4_000;
-    const positions = new Float32Array(count * 2);
-    for (let i = 0; i < count; i++) {
-      // Dense polyline far from the origin probe.
-      positions[i * 2] = 200 + (i % 100);
-      positions[i * 2 + 1] = 200 + Math.floor(i / 100);
-    }
-    // Short local stroke near (10,10) so nearest still hits something real.
-    positions[0] = 0;
-    positions[1] = 10;
-    positions[2] = 20;
-    positions[3] = 10;
-    const plotScene = scene();
-    plotScene.batches = [
-      {
-        kind: "paths",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions,
-        rowIndex: Uint32Array.from({ length: count }, (_, index) => index),
-        pathOffsets: new Uint32Array([0, count]),
-        strokes: [null],
-        linewidth: 2,
-        alpha: 1,
-        curve: "linear",
-      },
-    ];
-    let batchIndexReads = 0;
-    plotScene.batches = new Proxy(plotScene.batches, {
-      get(target, property, receiver): unknown {
-        if (typeof property === "string" && /^\d+$/.test(property)) batchIndexReads++;
-        return Reflect.get(target, property, receiver) as unknown;
-      },
-    });
-    const store = buildCandidateStore(plotScene, {
-      datum: () => ({ xValue: 1, yValue: 2, autoMode: "exact" }),
-    });
-    void store.x;
-    batchIndexReads = 0;
-
-    const hit = store.nearest(10, 10, { mode: "exact", maxDistance: 3 });
-    expect(hit).not.toBeNull();
-    expect(hit!.id).toBeLessThan(4);
-    // Linear extended scan would touch the batch once per far vertex (~count).
-    expect(batchIndexReads).toBeLessThan(200);
-    expect(batchIndexReads).toBeLessThan(count / 10);
-
-    batchIndexReads = 0;
-    expect(store.hitTest(10, 10)?.id).toBeLessThan(4);
-    expect(batchIndexReads).toBeLessThan(200);
-    expect(batchIndexReads).toBeLessThan(count / 10);
-  });
-
-  it("filled path still uses subpath AABB so interior probes hit", () => {
-    // Regression: stroke edge boxes must not replace fill containment shortlist.
-    const areaScene = scene();
-    areaScene.batches = [
-      {
-        kind: "paths",
-        layerIndex: 0,
-        panelIndex: 0,
-        // Large triangle; probe deep inside, far from all vertices.
-        positions: new Float32Array([0, 0, 200, 0, 100, 200]),
-        rowIndex: new Uint32Array([0, 1, 2]),
-        pathOffsets: new Uint32Array([0, 3]),
-        strokes: [null],
-        fills: [null],
-        closed: true,
-        linewidth: 0,
-        alpha: 1,
-        curve: "linear",
-      },
-    ];
-    const areaStore = buildCandidateStore(areaScene, {
-      datum: () => ({ xValue: 1, yValue: 2, autoMode: "exact" }),
-    });
-    expect(areaStore.nearest(100, 50, { mode: "exact", maxDistance: 0 })?.id).toBeDefined();
-    expect(areaStore.hitTest(100, 50)?.id).toBeDefined();
-  });
-
-  it("builds stores that include glyph batches without throwing", () => {
-    const plotScene = scene();
-    plotScene.batches = [
-      {
-        kind: "glyphs",
-        layerIndex: 0,
-        panelIndex: 0,
-        positions: new Float32Array([10, 20, 30, 40]),
-        rowIndex: new Uint32Array([0, 1]),
-        texts: ["a", "b"],
-        color: null,
-        size: 12,
-        anchor: "start",
-        alpha: 1,
-      },
-    ];
-    const store = buildCandidateStore(plotScene, {
-      datum: () => ({ xValue: 1, yValue: 2 }),
-    });
-    expect(store.size).toBe(2);
-    expect(store.nearest(10, 20, { mode: "exact", maxDistance: 0 })).toBeNull();
-  });
-});
